+++ modules/Food_Service/Students/ActivityReport.php
@@ -25,7 +25,10 @@ Search( 'student_id', $extra );
 if ( UserStudentID()
 	&& ! $_REQUEST['modfunc'] )
 {
-	$where = "AND STUDENT_ID='" . UserStudentID() . "' ";
+	// Note: relate student based on his ACCOUNT_ID as STUDENT_ID was NULL for Transactions.
+	$where = " AND ACCOUNT_ID=(SELECT ACCOUNT_ID
+		FROM food_service_student_accounts
+		WHERE STUDENT_ID='" . UserStudentID() . "') ";
 
 	if ( ! empty( $_REQUEST['type_select'] ) )
 	{
@@ -61,7 +64,8 @@ if ( UserStudentID()
 		WHERE SYEAR='" . UserSyear() . "'
 		AND fst.TIMESTAMP BETWEEN '" . $date . "' AND date '" . $date . "' +1
 		AND SCHOOL_ID='" . UserSchool() . "'" . $where . "
-		ORDER BY " . ( ! empty( $_REQUEST['by_name'] ) ? "FULL_NAME," : '' ) . "fst.TRANSACTION_ID DESC", [ 'DATE' => 'ProperDateTime', 'SHORT_NAME' => 'bump_count' ] );
+		ORDER BY " . ( ! empty( $_REQUEST['by_name'] ) ? "FULL_NAME," : '' ) . "fst.TRANSACTION_ID DESC",
+		[ 'DATE' => 'ProperDateTime', 'SHORT_NAME' => 'bump_count' ] );
 
 		foreach ( (array) $RET as $RET_key => $RET_val )
 		{
@@ -132,7 +136,8 @@ if ( UserStudentID()
 		WHERE SYEAR='" . UserSyear() . "'
 		AND fst.TIMESTAMP BETWEEN '" . $date . "' AND date '" . $date . "' +1
 		AND SCHOOL_ID='" . UserSchool() . "'" . $where . "
-		ORDER BY " . ( ! empty( $_REQUEST['by_name'] ) ? "FULL_NAME," : '' ) . "fst.TRANSACTION_ID DESC", [ 'DATE' => 'ProperDateTime', 'SHORT_NAME' => 'bump_count' ] );
+		ORDER BY " . ( ! empty( $_REQUEST['by_name'] ) ? "FULL_NAME," : '' ) . "fst.TRANSACTION_ID DESC",
+		[ 'DATE' => 'ProperDateTime', 'SHORT_NAME' => 'bump_count' ] );
 
 		$columns = [
 			'TRANSACTION_ID' => _( 'ID' ),