+++ CHANGES.md
@@ -13,6 +13,7 @@ Changes in 12.4.1
 - SQL limit Activity Report to selected student/user in ActivityReport.php
 - Deprecate Sort by Name checkbox: we now only have 1 student/user in the list in ActivityReport.php
 - SQL INSERT STUDENT_ID INTO food_service_transactions in Food_Service/Students/Transactions.php
+- SQL relate student based on his ACCOUNT_ID as STUDENT_ID was NULL for Transactions in ActivityReport.php
 
 Changes in 12.4
 ---------------